refactor: extract SQLAlchemy models into models.py
